BankingStage Refactor: Separate Consumer Module
